New feature: contest equivalence class detection.
This actually introduces a lot of new code. The first step is to
automatically infer contest bounding boxes. Then, we use tesseract
to extract the text from each contest. We compute the equivalence
class by comparing when contests have nearly identical text. Then,
when entering text for a contest, text for equal contests will be
automatically filled in.