Starting redesign to support multi-box contests.
In the case where a contest is too long to fit on one column, it
will sometimes split on to a second column. When computing contest
equality, the two contests should be merged first, and the ordering
should include the two contests.

This commit breaks candidate reordering. Postprocessing will need
to be updated to handle multi-box contests.